fix structure 30 08
